Remove traits that deal with assembly
