--- crates/rune/src/compile/v1/assemble/expr_assign.rs
@@ -1,102 +1,0 @@
-use crate::compile::v1::assemble::prelude::*;
-
-/// Compile an `.await` expression.
-impl Assemble for ast::ExprAssign {
-    fn assemble(&self, c: &mut Compiler<'_>, needs: Needs) -> CompileResult<Asm> {
-        let span = self.span();
-        log::trace!("ExprAssign => {:?}", c.q.sources.source(c.source_id, span));
-
-        let supported = match &self.lhs {
-            // <var> = <value>
-            ast::Expr::Path(path) if path.rest.is_empty() => {
-                self.rhs.assemble(c, Needs::Value)?.apply(c)?;
-
-                let segment = path
-                    .first
-                    .try_as_ident()
-                    .ok_or_else(|| CompileError::msg(path, "unsupported path"))?;
-                let ident = segment.resolve(&c.q.storage, c.q.sources)?;
-                let var = c.scopes.get_var(c.q.visitor, &*ident, c.source_id, span)?;
-                c.asm.push(Inst::Replace { offset: var.offset }, span);
-                true
-            }
-            // <expr>.<field> = <value>
-            ast::Expr::FieldAccess(field_access) => {
-                let span = field_access.span();
-
-                // field assignment
-                match &field_access.expr_field {
-                    ast::ExprField::Path(path) => {
-                        if let Some(ident) = path.try_as_ident() {
-                            let slot = ident.resolve(&c.q.storage, c.q.sources)?;
-                            let slot = c.q.unit.new_static_string(ident.span(), slot.as_ref())?;
-
-                            self.rhs.assemble(c, Needs::Value)?.apply(c)?;
-                            c.scopes.decl_anon(self.rhs.span())?;
-
-                            field_access.expr.assemble(c, Needs::Value)?.apply(c)?;
-                            c.scopes.decl_anon(span)?;
-
-                            c.asm.push(Inst::ObjectIndexSet { slot }, span);
-                            c.scopes.undecl_anon(span, 2)?;
-                            true
-                        } else {
-                            false
-                        }
-                    }
-                    ast::ExprField::LitNumber(field) => {
-                        let number = field.resolve(c.q.storage(), c.q.sources)?;
-                        let index = number.as_tuple_index().ok_or_else(|| {
-                            CompileError::new(
-                                span,
-                                CompileErrorKind::UnsupportedTupleIndex { number },
-                            )
-                        })?;
-
-                        self.rhs.assemble(c, Needs::Value)?.apply(c)?;
-                        c.scopes.decl_anon(self.rhs.span())?;
-
-                        field_access.expr.assemble(c, Needs::Value)?.apply(c)?;
-                        c.asm.push(Inst::TupleIndexSet { index }, span);
-                        c.scopes.undecl_anon(span, 1)?;
-                        true
-                    }
-                }
-            }
-            ast::Expr::Index(expr_index_get) => {
-                let span = expr_index_get.span();
-                log::trace!(
-                    "ExprIndexSet => {:?}",
-                    c.q.sources.source(c.source_id, span)
-                );
-
-                self.rhs.assemble(c, Needs::Value)?.apply(c)?;
-                c.scopes.decl_anon(span)?;
-
-                expr_index_get.target.assemble(c, Needs::Value)?.apply(c)?;
-                c.scopes.decl_anon(span)?;
-
-                expr_index_get.index.assemble(c, Needs::Value)?.apply(c)?;
-                c.scopes.decl_anon(span)?;
-
-                c.asm.push(Inst::IndexSet, span);
-                c.scopes.undecl_anon(span, 3)?;
-                true
-            }
-            _ => false,
-        };
-
-        if !supported {
-            return Err(CompileError::new(
-                span,
-                CompileErrorKind::UnsupportedAssignExpr,
-            ));
-        }
-
-        if needs.value() {
-            c.asm.push(Inst::unit(), span);
-        }
-
-        Ok(Asm::top(span))
-    }
-}
