--- crates/rune/src/compile/v1/assemble/prelude.rs
@@ -1,15 +1,0 @@
-pub(crate) use crate::ast;
-pub(crate) use crate::ast::{Span, Spanned};
-pub(crate) use crate::compile::v1::{
-    Asm, Assemble, AssembleClosure, AssembleConst, AssembleFn, Compiler, Loop, Needs,
-};
-pub(crate) use crate::compile::{
-    CaptureMeta, CompileError, CompileErrorKind, CompileResult, Item, MetaKind,
-};
-pub(crate) use crate::parse::{ParseErrorKind, Resolve};
-pub(crate) use crate::runtime::{
-    ConstValue, Inst, InstAddress, InstAssignOp, InstOp, InstRangeLimits, InstTarget, InstVariant,
-    PanicReason,
-};
-pub(crate) use crate::{Hash, Protocol};
-pub(crate) use std::convert::TryFrom;
